feat(ui/incident-v2): add changes for nested option
